biomc_pp

Struct PythonPostProcess

pub(crate) struct PythonPostProcess {
    pub(crate) inner: PostProcess,
}

A struct that wraps the PostProcess type for Python bindings.

The PythonPostProcess struct is designed to provide a Python interface for the PostProcess struct, enabling the interaction with Rustโs PostProcess type in Python.

Fields

inner: A reference to the inner PostProcess object, which holds the actual post-processing logic that can be used in Python.

Example

# Example usage in Python
post_process = PostProcess(name,root)

Fields
inner: PostProcess
Implementations

impl PythonPostProcess

pub(crate) fn new(folder: &str, root: Option<String>) -> PyResult<Self>

Creates a new instance of PythonPostProcess.

This function serves as a constructor for creating aPostProcess struct for use in Python. It takes a folder path and an optional root string as arguments to initialize the PostProcess object.

Arguments

folder (str): A path to the folder where the post-processing files or resources are located.

root (Option<String>): An optional root path used for additional processing logic, or None if not provided.

Returns

PyResult<Self>: On success, returns a PythonPostProcess instance wrapped in a PyResult. On failure, raises a PyValueError with an error message.

Example

post_process = PostProcess("path/to/folder", "optional/root")

Errors

If the PostProcess::new function fails (e.g., due to invalid paths or other internal errors), this function returns a PyValueError with the message "Error creating object".

pub(crate) fn get_property_names(&self) -> PyResult<Vec<String>>

pub(crate) fn v_liquid(&self, py: Python<'_>) -> Py<PyArray2<f64>>

pub(crate) fn time(&self) -> PyResult<&[f64]>

Gets the time data

This function provides access to the time data stored within the PostProcess object. It returns a slice of f64 values representing the time series used in the post-processing.

Returns

PyResult<&[f64]>: A reference to a slice of f64 values representing the time series.

Example

post_process = PostProcess("path/to/folder")
time_data = post_process.time

pub(crate) fn n_export(&self) -> PyResult<usize>

Gets the number of exports

This function retrieves the number of export events.

Returns

PyResult<usize>: The number of exports, represented as a usize value.

Example

post_process = PostProcess("path/to/folder")
num_exports = post_process.n_export()

pub(crate) fn max_n_export_bio(&self) -> usize

pub(crate) fn weight(&self, py: Python<'_>) -> Py<PyArray1<f64>>

pub(crate) fn get_spatial_average_concentration( &self, py: Python<'_>, species: usize, phase: Phase, ) -> Py<PyArray1<f64>>

pub(crate) fn get_variance_concentration( &self, py: Python<'_>, species: usize, phase: Phase, ) -> Py<PyArray1<f64>>

pub(crate) fn get_spatial_average_mtr( &self, py: Python<'_>, species: usize, ) -> Py<PyArray1<f64>>

pub(crate) fn get_concentrations( &self, py: Python<'_>, phase: Phase, ) -> Py<PyArray3<f64>>

pub(crate) fn get_time_average_concentration( &self, py: Python<'_>, species: usize, position: usize, phase: Phase, ) -> Py<PyArray1<f64>>

pub(crate) fn get_spatial_property( &self, py: Python<'_>, name: &str, ) -> Py<PyArray2<f64>>

pub(crate) fn get_biomass_concentration( &self, py: Python<'_>, ) -> Py<PyArray2<f64>>

pub(crate) fn get_growth_in_number(&self, py: Python<'_>) -> Py<PyArray1<f64>>

pub(crate) fn get_number_particle<'py>( &self, py: Python<'_>, ) -> Py<PyArray2<f64>>

pub(crate) fn get_probes(&self, py: Python<'_>) -> Py<PyArray1<f64>>

pub(crate) fn get_properties( &self, py: Python<'_>, key: &str, i_export: usize, ) -> Py<PyArray1<f64>>

pub(crate) fn get_population_mean( &self, key: &str, i_export: usize, ) -> PyResult<f64>

pub(crate) fn get_time_population_mean( &self, py: Python<'_>, key: &str, ) -> Py<PyArray1<f64>>

pub fn get_histogram( &self, py: Python<'_>, n_bins: usize, i_export: usize, key: &str, ) -> (Py<PyArray1<f64>>, Py<PyArray1<f64>>)

pub fn mu_direct(&self, py: Python<'_>) -> PyResult<Py<PyArray1<f64>>>

pub fn estimate( &self, etype: Estimator, key: &str, i_export: usize, ) -> PyResult<f64>

pub(crate) fn get_spatial_average_biomass_concentration( &self, py: Python<'_>, ) -> PyResult<Py<PyArray1<f64>>>

pub fn estimate_time( &self, py: Python<'_>, etype: Estimator, key: &str, ) -> PyResult<Py<PyArray1<f64>>>

pub fn get_csv_tallies(&self, py: Python<'_>) -> PyResult<String>

pub(crate) fn get_tallies(&self, py: Python<'_>) -> Py<PyArray2<f64>>